requirements.txt
torch
torchvision
scikit-learn
flask
pillow
